+++ Lib/fontbakery/specifications/googlefonts.py
@@ -99,7 +99,7 @@ expected_check_ids = [
       , 'com.google.fonts/check/metadata/listed_on_gfonts' # METADATA.pb: Fontfamily is listed on Google Fonts API?
       , 'com.google.fonts/check/metadata/unique_full_name_values' # METADATA.pb: check if fonts field only has unique "full_name" values.
       , 'com.google.fonts/check/metadata/unique_weight_style_pairs' # METADATA.pb: check if fonts field only contains unique style:weight pairs.
-      , 'com.google.fonts/check/085' # METADATA.pb license is "APACHE2", "UFL" or "OFL"?
+      , 'com.google.fonts/check/metadata/license' # METADATA.pb license is "APACHE2", "UFL" or "OFL"?
       , 'com.google.fonts/check/086' # METADATA.pb should contain at least "menu" and "latin" subsets.
       , 'com.google.fonts/check/087' # METADATA.pb subsets should be alphabetically ordered.
       , 'com.google.fonts/check/088' # METADATA.pb: Copyright notice is the same in all fonts?
@@ -1482,10 +1482,10 @@ def com_google_fonts_check_metadata_unique_weight_style_pairs(family_metadata):
 
 
 @check(
-  id = 'com.google.fonts/check/085',
+  id = 'com.google.fonts/check/metadata/license',
   conditions = ['family_metadata']
 )
-def com_google_fonts_check_085(family_metadata):
+def com_google_fonts_check_metadata_license(family_metadata):
   """METADATA.pb license is "APACHE2", "UFL" or "OFL"?"""
   licenses = ["APACHE2", "OFL", "UFL"]
   if family_metadata.license in licenses: